requirements.txt
requests
python-telegram-bot
SQLAlchemy
psycopg2-binary
yt-dlp
